rename bench to benchmark file to keep the consistency
