+++ tests/tt_metal/distributed/CMakeLists.txt
@@ -15,7 +15,7 @@ target_sources(
         test_mesh_workload.cpp
         test_mesh_socket.cpp
         test_hd_sockets.cpp
-        bench_hd_sockets.cpp
+        benchmark_hd_sockets.cpp
         test_mesh_sub_device.cpp
         test_mesh_allocator.cpp
         test_mesh_events.cpp
